Test async preprocess and root print
