+++ tests/config/prettier-plugins/prettier-plugin-async-printer/package.json
@@ -1,6 +1,6 @@
 {
   "private": true,
-  "name": "prettier-plugin-async-parser",
+  "name": "prettier-plugin-async-printer",
   "version": "1.0.0",
   "main": "./index.cjs"
 }
