--- tests/format/misc/plugins/async-parser/jsfmt.spec.js
@@ -1,9 +1,0 @@
-import createEsmUtils from "esm-utils";
-
-const { require } = createEsmUtils(import.meta);
-
-const plugins = [
-  require("../../../../config/prettier-plugins/prettier-plugin-async-parser/index.cjs"),
-];
-
-run_spec(import.meta, ["async-parser"], { plugins });
