+++ tests/format/misc/plugins/async-printer/async-printer.txt
@@ -1,0 +1,1 @@
+Lorem Ipsum
