--- tests/format/misc/plugins/embed-async-parser/with-aync-parser-code-block.md
@@ -1,7 +1,0 @@
-```async-parser
-Async parsers can be used in embeds, so this text will be replaced.
-```
-
-```uppercase-rocks
-This text should be uppercased.
-```
