+++ tests/format/misc/plugins/embed-async-printer/jsfmt.spec.js
@@ -1,6 +1,6 @@
 const plugins = await Promise.all(
   [
-    "../../../../config/prettier-plugins/prettier-plugin-async-parser/index.cjs",
+    "../../../../config/prettier-plugins/prettier-plugin-async-printer/index.cjs",
     "../../../../config/prettier-plugins/prettier-plugin-uppercase-rocks/index.js",
   ].map(async (plugin) => (await import(plugin)).default)
 );
