+++ tests/format/misc/plugins/embed-async-printer/with-aync-printer-code-block.md
@@ -1,0 +1,7 @@
+```async-printer
+This text will be lowercased and prefixed with a comment.
+```
+
+```uppercase-rocks
+This text should be uppercased.
+```
